Fix compilation error '<symbol> changed binding to STB_GLOBAL'
